managed_bytes: introduce with_linearized()
A large managed_bytes blob can be scattered in lsa memory.  Usually this is
fine, but someone we want to examine it in place without copying it out, but
using contiguous iterators for efficiency.

For this use case, introduce with_linearized(const managed_bytes&, Func),
which runs a function on the data in the blob.  If the data happens to be
contiguous (which is the common case), it is passed to the function without
change.  If it is scattered, then we temporarily gather it and pass the copy
to the function.