lsa: provide an API for variable-size allocation
Instead of specifying a fixed size, allow the caller to specify a size
range.  This lets lsa select a size that will fit available space in the
segment - a few bytes if we have an almost-full segment, or an entire segment
if we just opened a new one.  This allows a variable-size-aware caller to
get occupancy of >99%.