Convert ring_position and token ranges to be nonwrapping
Wrapping ranges are a pain, so we are moving wrap handling to the edges.

Since cql can't generate wrapping ranges, this means thrift and the ring
maintenance code; also range->ring transformations need to merge the first
and last ranges.

Note: this patch goes too far by converting IDL as well.  IDL should be left
alone and the server code should be converted to unwrap, so that old->new
streaming can work.