streaming: add synchronous streaming interface
The synchronous interface replies only on rpc verbs that do not
complete until all they were supposed to do is done, so if the remote
fails, they are not left with incomplete operations. Because of that,
they do not rely on timeouts. On the other hand, an RPC connection failure
will fail the entire streaming operation, so use with moderately sized
streams to avoid losing too much work.

Only a single direction is defined so far: server->client. The opposite
direction will be implemented on top.

The interface talks to the rest of the system via abstract interfaces
to avoid dependency problems.