sstables: introduce new type for shared_sstable
lw_shared_ptr<sstable> is extensively used, but has the downside of
requiring the definition of class sstable to be visible in all its
glory. The root cause is that the inline lw_shared_ptr destructor needs
to know whether sstable has an overloaded delete operator. This increases
compile time and recompilation due to dependencies.

To fix, introduce a new class for shared_sstable that does the same
basic reference counting but does not depend on sstable itself. Any
access to the sstable being referenced will need the class definition,
but just moving shared_sstable objects around does not, and those
actions are still inlineable.