cql: select_statement: make execute() and execute_internal() equivalent
execute_internal(), for some code paths, differs from execute by the
following:
 1. it uses CL_ONE unconditionally
 2. it has no query timeout
 3. it doesn't use execution stages

for other code paths, it just calls execute.

As preparation for getting rid of execute_internal(), unify the two
code paths. execute() will now look at whether the query is internal
or not and override the timeout and consistency level appropriately.
It will now always use the execution stage, but that has no semantic
effect.

The consistency level and timeout should properly be provided by the
caller, but that is left to a later change.