build: add relocatable package target
A relocatable package contains the Scylla (and iotune)
executables (in a bin/ directory), any libraries they may need (lib/)
the configuration file defaults (conf/) and supporting scripts (dist/).
The libraries are picked up from the host; libc is not picked up since
it is tied up strongly to the dynamic loader (ld-linux.so).

With these files, it is possible to run a fully functional Scylla
instance on any Linux distribution, provided that its glibc is
new enough. To satisfy that requirement, build on a sufficiently
old host.

The packages are created by running

    ninja build/release/scylla-package.tar

or

    ninja build/debug/scylla-package.tar