database: multiple tenant configuration
Instead of having just a single scheduling_group for statements,
allow different scheduling_groups for different roles. A default
tenant is provided as a fallback, in case the user has no scheduling
group associated with the role.

There is no mechanism yet for breaking out of the default group;
that will be provided in a following patch.