dht: change token class to use fixed-size storage
Now that murmur3_partitioner is our only partitoiner, we can use
fixed-size storage for the token, saving the space and time overhead
of dealing with variable-size storage.

We create a new type token_data that is serialized in the same way
as variable-size data (with the size prefix) to preserve compatibility.

Tests: unit (release).