tests: flat_reader_assertions: drain reader on close
An undrained reader will be close in the background, and if this is not
done quickly enough, it can happen after the test terminates and this
gets reported as a memory leak.

Fix by draining in the foreground.

We should consider a flat_mutation_reader::close() method, or a resource
manager for flat mutation readers, for the general case of uncontrolled
background operations due to short reads.