table: stop writes if at risk of running out of memory on reads
While compaction is supposed to keep the number of sstables under control,
we cannot guarantee that this will happen. As the number of sstables grows,
so does the amount of memory needed to service a read. To prevent out-of-memory
events, stop creating new sstables (via table::flush()) until compaction can
catch up. This is akin to the hard limit on dirty memory; we prefer that the
controllers will apply back pressure and reduce write rate, but if even that
doesn't work then a hard stop is better than a crash.

We implement this by adding a condition variable, and waiting on that condition
variable until we compaction strategy agrees that a new sstable is not dangeous.