sstable_set: add a method to estimate the number of sstables that will participate in a read
If a read accesses too many sstables, we can run out of memory (for buffers) while
processing the read. To avoid this, we'd like to stop creating new sstables if that
will take us over the limit. To know when to stop creating new sstables, we need
to estimate how many sstables a read will currently require.

This patch implements the first step: a way for sstable_set to report the number
of sstables that will participate in a read. Since the sstables used in a read are
determinted by sstable_set::select() and sstable_set::incremental_selector, this
is the entry point for estimation.

This patch adds sstable_set::current_estimated_sstables_for_read() to provide
that estimation, and implements it for the two types of sstable sets we support.