config: use non-shared pointer to pass extensions
We will soon make config a per-shard object, so sharing the extensions across shards
will be made more difficult. Simplify by using a naked pointer instead of a shared
pointer. This requires callers to keep track of extensions lifetime, but that's
essentially the lifetime of the application, so it's not hard.