toppartitions: fix race between listener removal and reads
Data listener reads are implemented as flat_mutation_readers, which
take a reference to the listener and then execute asynchronously.
The listener can be removed between the time when the reference is
taken and actual execution, resulting in a dangling pointer
dereference.

Fix by using a gate to add reference counting. Note that writes
don't need protection because they execute atomically.

Fixes #4661.

Tests: unit (dev)