cql3: select_statement: limit concurrency of indexed select row reads
Currently, we read all of the rows of the proposed result page concurrently. This
requires a lot of memory, generates stalls, and causes the reader_concurrency_semaphore
(when moved in front of the cache) to overflow.

Fix by limiting concurrency to 100.

A better fix would involve exponential increase with cutoff at the target page
size (in bytes, not rows), but that is left for later.

Ref #3807.