table: convery query() to a coroutine
Simplify the code by converting it to a coroutine. This is a
mechanical conversion:

    then -> co_await
    do_until -> while
    finally -> defer()

Indentation is left inconsistent to ease review; will be fixed
in the following patch.

This is a slight pessimization in performance, since the coroutine
always allocates, while the non-coroutine code could elide the
allocation in many cache hit scenarios, but a later patch recovers
the performance loss by removing another allocation.