reconcilable_result_builder: warn/fail if too many tombstones are encountered
If too many tombstones are encountered in a run,
reconcilable_result_builder will accumulate them, disregarding any
memory limit, since we must provide at least one row (and we don't
support breaking a page after a dead row). This can consume unbounded
amounts of memory if a partition full of row tombstones and no live
rows is encountered.

Prepare to prevent this by issuing a warning when a soft limit
is encountered, and failing the query when a hard limit is
encountered. The reconcilable_result_builder configuration is now
defaulted to infinite, this will be adjusted at caller sites in the
next patch.