utils: uuid: add namespace-scope ostream streaming operator declaration
I see a miscompile on aarch64 where clang resolves the call to
operator<<(std::ostream, utils::UUID) to constant -1, deep in the
bowels of fmt. This is fixed by supplying a namespace-scope
declaration to complement the class-scope friend declaration.

This is clearly a bug in Clang, but very difficult to narrow down.
I previously observed something similar on x86, though I can't recall
exactly where.