cql3: statement_restrictions: wrap get_partition_key_ranges
statement_restrictions::get_partition_key_ranges() re-interprets
the expressions used to specify the partition key. This means that
the analysis phase (determining what those expressions are and how
they are to be used) and the execution phase (using them) are in separate
places. This makes it very hard to refactor while preserving correctness.

As a first step in unifying the two phases, we move the selection
of the strategy (using token, cartesian product, or single partition)
happen at analysis time, by making the if-tree return a function to
be executed at execution time, rather than running the if-tree itself
at execution time.