cql3: statement_restrictions: do not collect subscripted partition key columns
An indexed SELECT of the from

SELECT ...
WHERE pk['sub'] = ?

is impossible because our indexes do not support frozen maps, and
partition key collections must be frozen. Stop collecting such constructs
for the purpose of determining the partition range. This reduces having
to deal with combinations of restrictions on the column and its entries
later on.

In case we start supporting indexes on frozen maps, leave an
on_internal_error to remind us.