cql3: statement_restrictions: pre-analyze token range restrictions
Convert token range restrictions to the analyzed_column format we
introduced earlier, where we have a function to solve for the token
range rather than running the analysis at runtime. Again the truth is
that the function will delegate to possible_partition_token_values()
which actually will do the analysis at runtime, but it's one step closer.