utils: config_file: don't use extern fully specialized variable templates
Declaring-but-not-defining a fully specialized template is a great way to
cut dependencies between users and providers, but unfortunately not
supported for variable templates. Clang 18 does support it, but
apparently it is a misinterpretation of the standard, and was removed
in clang 19.

We started using this non-feature in 7ed89266b30f782c05.

Fortunately, there is a simple workaround: instead of a variable
template, use a class template with a static variable. This provides
a declaration for the static variable. Definitions can then be provided
out-of-line as before.