test: pylib: log_browsing: use direct deadline check for timeout instead of asyncio.timeout
wait_for() uses asyncio.timeout() to time out. We observed Python
running in a tight loop here and not timing out (#27860); it's possible that
the async timeout doesn't work well with the `universalasync` library
that is used here to turn asynchronous code into synchronous.

Attempt to fix it (if it is indeed the problem) by using a direct deadline
check. Since the other coroutines are expected to terminate quickly (they
only read a line, so not network-dependent or similar), the only thing
that can take any amount of time is the sleep. Checking just before the
sleep is sufficient.