test: pylib: log_browsing: fix tight loop on unterminated line
wait_for() accumulates a log line in a variable named `line`.
If the log terminates in a line without a line terminator, the
variable `line` will never be cleared, and the `else: sleep()`
branch will never be called, causing a tight loop. It's also
possible the tight loop won't allow the outer `with asyncio.timeout()`
to take effect, as was observed in #27860, though this wasn't proven.

Fix by always taking the else: branch if we don't have a full line.
There's no point in matching the patterns against an incomplete line.